fix: use int64 dtype for scalar constants in shape propagation
Was producing int32, inconsistent with int64_t used throughout the
C++ runtime (tensor shapes, Params.ints, kernel indices).

+++ py/fxfusion/passes/shape_prop.py
@@ -117,7 +117,7 @@ class ShapePropPass:
         elif isinstance(result, int):
             scalar = torch.tensor(
                 result,
-                dtype=torch.int32,
+                dtype=torch.int64,
                 device=scalar_device_from_inputs(),
             )
